+++ rich/protocol.py
@@ -1,5 +1,4 @@
 from typing import Any, cast, Set, TYPE_CHECKING
-from inspect import isclass
 
 if TYPE_CHECKING:
     from rich.console import RenderableType
@@ -28,7 +27,7 @@ def rich_cast(renderable: object) -> "RenderableType":
     from rich.console import RenderableType
 
     rich_visited_set: Set[type] = set()  # Prevent potential infinite loop
-    while hasattr(renderable, "__rich__") and not isclass(renderable):
+    while hasattr(renderable, "__rich__") and not isinstance(renderable, type):
         # Detect object which claim to have all the attributes
         if hasattr(renderable, _GIBBERISH):
             return repr(renderable)