+++ rich/segment.py
@@ -1,7 +1,6 @@
 from enum import IntEnum
 from functools import lru_cache
 from itertools import filterfalse
-from logging import getLogger
 from operator import attrgetter
 from typing import (
     TYPE_CHECKING,
@@ -28,8 +27,6 @@ from .style import Style
 
 if TYPE_CHECKING:
     from .console import Console, ConsoleOptions, RenderResult
-
-log = getLogger("rich")
 
 
 class ControlType(IntEnum):
@@ -274,6 +271,37 @@ class Segment(NamedTuple):
                 append(segment)
         if line:
             yield line
+
+    @classmethod
+    def split_lines_terminator(
+        cls, segments: Iterable["Segment"]
+    ) -> Iterable[Tuple[List["Segment"], bool]]:
+        """Split a sequence of segments in to a list of lines and a boolean to indicate if there was a new line.
+
+        Args:
+            segments (Iterable[Segment]): Segments potentially containing line feeds.
+
+        Yields:
+            Iterable[List[Segment]]: Iterable of segment lists, one per line.
+        """
+        line: List[Segment] = []
+        append = line.append
+
+        for segment in segments:
+            if "\n" in segment.text and not segment.control:
+                text, style, _ = segment
+                while text:
+                    _text, new_line, text = text.partition("\n")
+                    if _text:
+                        append(cls(_text, style))
+                    if new_line:
+                        yield (line, True)
+                        line = []
+                        append = line.append
+            else:
+                append(segment)
+        if line:
+            yield (line, False)
 
     @classmethod
     def split_and_crop_lines(